Fixed T5 double tokenization, moved token parsing to shared trait function
