+++ main/src/preprocessing/vocab/sentence_piece_vocab.rs
@@ -20,7 +20,7 @@ use crate::Vocab;
 use std::collections::HashMap;
 use crate::preprocessing::vocab::base_vocab::swap_key_values;
 use std::process;
-use crate::preprocessing::tokenizer::base_tokenizer::{TokenRef, OffsetSize};
+use crate::preprocessing::tokenizer::base_tokenizer::{TokenRef, OffsetSize, Token, Offset};
 use crate::preprocessing::vocab::sentencepiece_proto::sentencepiece_model::ModelProto;
 
 
@@ -186,6 +186,38 @@ impl SentencePieceModel {
         };
         best_sequence.reverse();
         best_sequence
+    }
+
+    pub fn parse_nodes_to_tokens(&self, nodes: Vec<&Node>) -> Vec<Token> {
+        let mut output: Vec<Token> = Vec::with_capacity(nodes.len() + 1);
+        let mut is_prev_unknown = false;
+        for node in nodes {
+            // Group unknown tokens
+            if is_prev_unknown & (node.index == 0) {
+                let prev_token = output.last().unwrap();
+                let mut text = prev_token.text.clone();
+                text.push_str(node.text);
+                let mut reference_offsets = prev_token.reference_offsets.clone();
+                reference_offsets.extend_from_slice(node.reference_offsets);
+                let consolidated_unknown = Token {
+                    text,
+                    offset: Offset { begin: 0, end: 0 },
+                    reference_offsets,
+                    mask: Default::default(),
+                };
+                output.pop();
+                output.push(consolidated_unknown);
+            } else {
+                output.push(Token {
+                    text: node.text.to_owned(),
+                    offset: Offset { begin: 0, end: 0 },
+                    reference_offsets: node.reference_offsets.to_vec(),
+                    mask: Default::default(),
+                });
+            }
+            is_prev_unknown = node.index == 0;
+        }
+        output
     }
 }
